+++ lib/docs/CHANGELOG.md
@@ -92,6 +92,8 @@ Added:
 - Use dedicated tile when Morgoth loses his crown (fixes #170)
   ([#172](https://github.com/sil-quirk/sil-q/pull/172))
   - thanks @MicroChasm for the suggestion
+- Depleted light sources use depleted version of their tile (fixes #163)
+- Change tile design of The Ring of Barahir to match the lore (fixes #181)
 
 Changed: